Move data functions to overlay.py
